Refactor: Extract withUserPgPool
